requirements.txt
celery==4.2.1
dateutils==0.6.6
Flask==1.0.2
Flask-Celery==2.4.3
Flask-Cors==3.0.6
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
flasgger==0.9.0
sqlalchemy==1.2.10